feat: Add exit option to trusted command rule creation Add a new option 'Exit rule creation and don't run any commands' to the interactive trusted command rule creation menu. This allows users to cancel both rule creation and command execution when prompted with the 'c' option. Changes: - Add option 5 to the interactive menu display - Update choice range from 1-4 to 1-5 in user prompt - Handle new exit option by returning to PromptUser state - Maintain existing behavior for options 1-4 (create rules or run command) The new option provides users with a clean way to exit the rule creation process without executing the command or creating any trusted command rules. Refactor duplicate code.
